--aded animation between routes -- added comments -- lazy loaded images
