+++ src/app/app.module.ts
@@ -1,25 +1,21 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule }  from '@angular/forms';
-import { HttpModule } from '@angular/http';
-
-import { LocationStrategy, PathLocationStrategy} from '@angular/common';
-
-import { AppComponent } from './app.component';
-import { routing, appRoutingProviders }  from './app.routing';
-import { HttpService }  from './services/http.service';
-
-import { PageComponent } from './page/page.component';
-import { AboutComponent }  from './about/about.component';
-import { ErrorComponent } from './error/error.component';
+import {NgModule} from "@angular/core";
+import {BrowserModule} from "@angular/platform-browser";
+import {FormsModule} from "@angular/forms";
+import {HttpModule} from "@angular/http";
+import {LocationStrategy, PathLocationStrategy} from "@angular/common";
+import {AppComponent} from "./components/app/app.component";
+import {routing, appRoutingProviders} from "./app.routing";
+import {HttpService} from "./services/http.service";
+import {PageComponent} from "./components/page/page.component";
+import {AboutComponent} from "./components/about/about.component";
+import {ErrorComponent} from "./components/error/error.component";
 import {SeoService} from "./services/SeoService";
 import {DataParseService} from "./services/DataParseService";
 import {MenuItemComponent} from "./shared/menu/menu.item.component";
 import {MenuComponent} from "./shared/menu/menu.component";
 import {RouteController} from "./services/RouteController";
-import {ContentComponent} from "./content/content.component";
+import {ContentComponent} from "./components/content/content.component";
+import {DataSrc} from "./diractives/DataSrc/DataSrc";
 
 @NgModule({
   imports: [
@@ -35,7 +31,8 @@ import {ContentComponent} from "./content/content.component";
     ErrorComponent,
     MenuComponent,
     MenuItemComponent,
-    ContentComponent
+    ContentComponent,
+    DataSrc
   ],
   providers: [
     appRoutingProviders,
