+++ src/app/components/page/page.component.html
@@ -1,0 +1,14 @@
+<div [@wrapper]="state">
+  <div *ngIf="loaded">
+    <h2>{{umbpage?.name}}</h2>
+    <div [innerHTML]="umbpage?.contentBody"></div>
+    <!--
+    <img src="{{umbpage?.contentImages[0].croppedUrl}}" alt="">
+    -->
+
+    <!-- fake image -->
+    <img *ngFor="let img of imgs" [datasrc]="img.croppedUrl" alt={{img?.altText}}>
+  </div>
+  <div class="loader centered" *ngIf="!loaded"><div class="loader-inner ball-scale"><div></div></div></div>
+
+</div>
