+++ src/app/constants/pages.ts
@@ -1,6 +1,6 @@
 
 
-import {PageComponent} from "../page/page.component";
+import {PageComponent} from "../components/page/page.component";
 
 export const PAGE_ALIAS={
   "PageComponent": PageComponent
