+++ src/index.html
@@ -4,11 +4,11 @@
   <base href="/">
   <meta charset="utf-8">
   <title>Webapp</title>
-  
+
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
-
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
 </head>
 <body>
   <app-root>
